chore: access media through devserver app
