+++ imports/plugins/core/graphql/server/index.js
@@ -1,12 +1,13 @@
 import { Meteor } from "meteor/meteor";
 import { MongoInternals } from "meteor/mongo";
 import { WebApp } from "meteor/webapp";
+import { NoMeteorMedia } from "/imports/plugins/core/files/server";
 import createApolloServer from "./createApolloServer";
 import defineCollections from "./defineCollections";
 import methods from "./methods";
 import queries from "./queries";
 
-const collections = {};
+const collections = { Media: NoMeteorMedia };
 
 const { db } = MongoInternals.defaultRemoteCollectionDriver().mongo;
 defineCollections(db, collections);
