+++ package.json
@@ -20,8 +20,8 @@
   },
   "dependencies": {
     "@babel/runtime": "7.0.0-beta.38",
-    "@reactioncommerce/file-collections": "0.4.5",
-    "@reactioncommerce/file-collections-sa-gridfs": "0.0.1",
+    "@reactioncommerce/file-collections": "0.5.0",
+    "@reactioncommerce/file-collections-sa-gridfs": "0.0.2",
     "@reactioncommerce/hooks": "1.0.2",
     "@reactioncommerce/job-queue": "1.0.4",
     "@reactioncommerce/logger": "1.1.1",
